adding abstract classes for common properties of objects
