+++ src/main/java/com/hashmapinc/tempus/WitsmlObjects/common/CsDocumentInfo.java
@@ -1,0 +1,110 @@
+package com.hashmapinc.tempus.WitsmlObjects.common;
+
+import javax.xml.bind.annotation.*;
+import javax.xml.datatype.XMLGregorianCalendar;
+
+@XmlTransient
+public abstract class CsDocumentInfo {
+
+    @XmlSchemaType(name = "dateTime")
+    protected XMLGregorianCalendar documentDate;
+    protected String disclaimer;
+    protected String owner;
+    protected String comment;
+
+    /**
+     * Gets the value of the documentDate property.
+     *
+     * @return
+     *     possible object is
+     *     {@link XMLGregorianCalendar }
+     *
+     */
+    public XMLGregorianCalendar getDocumentDate() {
+        return documentDate;
+    }
+
+    /**
+     * Sets the value of the documentDate property.
+     *
+     * @param value
+     *     allowed object is
+     *     {@link XMLGregorianCalendar }
+     *
+     */
+    public void setDocumentDate(XMLGregorianCalendar value) {
+        this.documentDate = value;
+    }
+
+    /**
+     * Gets the value of the disclaimer property.
+     *
+     * @return
+     *     possible object is
+     *     {@link String }
+     *
+     */
+    public String getDisclaimer() {
+        return disclaimer;
+    }
+
+    /**
+     * Sets the value of the disclaimer property.
+     *
+     * @param value
+     *     allowed object is
+     *     {@link String }
+     *
+     */
+    public void setDisclaimer(String value) {
+        this.disclaimer = value;
+    }
+
+    /**
+     * Gets the value of the owner property.
+     *
+     * @return
+     *     possible object is
+     *     {@link String }
+     *
+     */
+    public String getOwner() {
+        return owner;
+    }
+
+    /**
+     * Sets the value of the owner property.
+     *
+     * @param value
+     *     allowed object is
+     *     {@link String }
+     *
+     */
+    public void setOwner(String value) {
+        this.owner = value;
+    }
+
+    /**
+     * Gets the value of the comment property.
+     *
+     * @return
+     *     possible object is
+     *     {@link String }
+     *
+     */
+    public String getComment() {
+        return comment;
+    }
+
+    /**
+     * Sets the value of the comment property.
+     *
+     * @param value
+     *     allowed object is
+     *     {@link String }
+     *
+     */
+    public void setComment(String value) {
+        this.comment = value;
+    }
+}
