+++ src/main/java/com/hashmapinc/tempus/WitsmlObjects/v1311/AbstractMeasure.java
@@ -4,7 +4,6 @@ import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlSeeAlso;
 import javax.xml.bind.annotation.XmlType;
-import javax.xml.bind.annotation.XmlValue;
 
 /**
  * The intended abstract supertype of all quantities that have a value 
@@ -78,26 +77,6 @@ import javax.xml.bind.annotation.XmlValue;
     VolumeFlowRateMeasure.class,
     MagneticInductionMeasure.class
 })
-public abstract class AbstractMeasure {
-
-    @XmlValue
-    protected double value;
-
-    /**
-     * This type disallows an "empty" double value.
-     * 			This type should not be used directly except to derive another type.
-     * @return returns the value as a Double
-     */
-    public double getValue() {
-        return value;
-    }
-
-    /**
-     * Sets the value of the value property.
-     * @param value The value to set as a Double
-     */
-    public void setValue(double value) {
-        this.value = value;
-    }
+public abstract class AbstractMeasure extends com.hashmapinc.tempus.WitsmlObjects.common.AbstractMeasure {
 
 }
