+++ src/main/java/com/hashmapinc/tempus/WitsmlObjects/v1311/CsCommonData.java
@@ -2,9 +2,7 @@ package com.hashmapinc.tempus.WitsmlObjects.v1311;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlSchemaType;
 import javax.xml.bind.annotation.XmlType;
-import javax.xml.datatype.XMLGregorianCalendar;
 
 
 /**
@@ -40,88 +38,9 @@ import javax.xml.datatype.XMLGregorianCalendar;
     "itemState",
     "comments"
 })
-public class CsCommonData {
+public class CsCommonData extends com.hashmapinc.tempus.WitsmlObjects.common.CsCommonData {
 
-    protected String sourceName;
-    @XmlSchemaType(name = "dateTime")
-    protected XMLGregorianCalendar dTimCreation;
-    @XmlSchemaType(name = "dateTime")
-    protected XMLGregorianCalendar dTimLastChange;
-    @XmlSchemaType(name = "string")
     protected ItemState itemState;
-    protected String comments;
-
-    /**
-     * Gets the value of the sourceName property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getSourceName() {
-        return sourceName;
-    }
-
-    /**
-     * Sets the value of the sourceName property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setSourceName(String value) {
-        this.sourceName = value;
-    }
-
-    /**
-     * Gets the value of the dTimCreation property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link XMLGregorianCalendar }
-     *     
-     */
-    public XMLGregorianCalendar getDTimCreation() {
-        return dTimCreation;
-    }
-
-    /**
-     * Sets the value of the dTimCreation property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link XMLGregorianCalendar }
-     *     
-     */
-    public void setDTimCreation(XMLGregorianCalendar value) {
-        this.dTimCreation = value;
-    }
-
-    /**
-     * Gets the value of the dTimLastChange property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link XMLGregorianCalendar }
-     *     
-     */
-    public XMLGregorianCalendar getDTimLastChange() {
-        return dTimLastChange;
-    }
-
-    /**
-     * Sets the value of the dTimLastChange property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link XMLGregorianCalendar }
-     *     
-     */
-    public void setDTimLastChange(XMLGregorianCalendar value) {
-        this.dTimLastChange = value;
-    }
 
     /**
      * Gets the value of the itemState property.
@@ -146,29 +65,4 @@ public class CsCommonData {
     public void setItemState(ItemState value) {
         this.itemState = value;
     }
-
-    /**
-     * Gets the value of the comments property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getComments() {
-        return comments;
-    }
-
-    /**
-     * Sets the value of the comments property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setComments(String value) {
-        this.comments = value;
-    }
-
 }
