+++ src/main/java/com/hashmapinc/tempus/WitsmlObjects/v1311/GenericMeasure.java
@@ -2,7 +2,6 @@ package com.hashmapinc.tempus.WitsmlObjects.v1311;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlAttribute;
 import javax.xml.bind.annotation.XmlType;
 
 
@@ -30,34 +29,7 @@ import javax.xml.bind.annotation.XmlType;
 @XmlAccessorType(XmlAccessType.FIELD)
 @XmlType(name = "genericMeasure")
 public class GenericMeasure
-    extends AbstractMeasure
+    extends com.hashmapinc.tempus.WitsmlObjects.common.GenericMeasure
 {
-
-    @XmlAttribute(name = "uom")
-    protected String uom;
-
-    /**
-     * Gets the value of the uom property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getUom() {
-        return uom;
-    }
-
-    /**
-     * Sets the value of the uom property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setUom(String value) {
-        this.uom = value;
-    }
 
 }
