+++ src/main/java/com/hashmapinc/tempus/WitsmlObjects/v1411/CsCommonData.java
@@ -12,9 +12,7 @@ import java.util.ArrayList;
 import java.util.List;
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlSchemaType;
 import javax.xml.bind.annotation.XmlType;
-import javax.xml.datatype.XMLGregorianCalendar;
 
 
 /**
@@ -62,94 +60,15 @@ import javax.xml.datatype.XMLGregorianCalendar;
     "extensionAny",
     "extensionNameValue"
 })
-public class CsCommonData {
+public class CsCommonData extends com.hashmapinc.tempus.WitsmlObjects.common.CsCommonData {
 
-    protected String sourceName;
-    @XmlSchemaType(name = "dateTime")
-    protected XMLGregorianCalendar dTimCreation;
-    @XmlSchemaType(name = "dateTime")
-    protected XMLGregorianCalendar dTimLastChange;
-    @XmlSchemaType(name = "string")
     protected ItemState itemState;
     protected String serviceCategory;
-    protected String comments;
     protected List<TimestampedTimeZone> acquisitionTimeZone;
     protected RefNameString defaultDatum;
     protected Boolean privateGroupOnly;
     protected CsExtensionAny extensionAny;
     protected List<CsExtensionNameValue> extensionNameValue;
-
-    /**
-     * Gets the value of the sourceName property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getSourceName() {
-        return sourceName;
-    }
-
-    /**
-     * Sets the value of the sourceName property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setSourceName(String value) {
-        this.sourceName = value;
-    }
-
-    /**
-     * Gets the value of the dTimCreation property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link XMLGregorianCalendar }
-     *     
-     */
-    public XMLGregorianCalendar getDTimCreation() {
-        return dTimCreation;
-    }
-
-    /**
-     * Sets the value of the dTimCreation property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link XMLGregorianCalendar }
-     *     
-     */
-    public void setDTimCreation(XMLGregorianCalendar value) {
-        this.dTimCreation = value;
-    }
-
-    /**
-     * Gets the value of the dTimLastChange property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link XMLGregorianCalendar }
-     *     
-     */
-    public XMLGregorianCalendar getDTimLastChange() {
-        return dTimLastChange;
-    }
-
-    /**
-     * Sets the value of the dTimLastChange property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link XMLGregorianCalendar }
-     *     
-     */
-    public void setDTimLastChange(XMLGregorianCalendar value) {
-        this.dTimLastChange = value;
-    }
 
     /**
      * Gets the value of the itemState property.
@@ -197,30 +116,6 @@ public class CsCommonData {
      */
     public void setServiceCategory(String value) {
         this.serviceCategory = value;
-    }
-
-    /**
-     * Gets the value of the comments property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getComments() {
-        return comments;
-    }
-
-    /**
-     * Sets the value of the comments property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setComments(String value) {
-        this.comments = value;
     }
 
     /**
