+++ src/main/java/com/hashmapinc/tempus/WitsmlObjects/v1411/GenericMeasure.java
@@ -10,7 +10,6 @@ package com.hashmapinc.tempus.WitsmlObjects.v1411;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlAttribute;
 import javax.xml.bind.annotation.XmlType;
 
 
@@ -38,34 +37,7 @@ import javax.xml.bind.annotation.XmlType;
 @XmlAccessorType(XmlAccessType.FIELD)
 @XmlType(name = "genericMeasure")
 public class GenericMeasure
-    extends AbstractMeasure
+    extends com.hashmapinc.tempus.WitsmlObjects.common.GenericMeasure
 {
-
-    @XmlAttribute(name = "uom", required = true)
-    protected String uom;
-
-    /**
-     * Gets the value of the uom property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getUom() {
-        return uom;
-    }
-
-    /**
-     * Sets the value of the uom property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setUom(String value) {
-        this.uom = value;
-    }
 
 }
