+++ src/main/java/com/hashmapinc/tempus/WitsmlObjects/v1411/ObjWellbore.java
@@ -10,12 +10,8 @@ package com.hashmapinc.tempus.WitsmlObjects.v1411;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlAttribute;
-import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlSchemaType;
 import javax.xml.bind.annotation.XmlType;
-import javax.xml.datatype.XMLGregorianCalendar;
-
 
 /**
  * <p>Java class for obj_wellbore complex type.
@@ -71,16 +67,9 @@ import javax.xml.datatype.XMLGregorianCalendar;
     "commonData",
     "customData"
 })
-public class ObjWellbore {
+public class ObjWellbore extends com.hashmapinc.tempus.WitsmlObjects.common.ObjWellbore {
 
-    @XmlElement(required = true)
-    protected String nameWell;
-    @XmlElement(required = true)
-    protected String name;
     protected RefNameString parentWellbore;
-    protected String number;
-    protected String suffixAPI;
-    protected String numGovt;
     @XmlSchemaType(name = "string")
     protected WellStatus statusWellbore;
     protected Boolean isActive;
@@ -90,9 +79,6 @@ public class ObjWellbore {
     protected WellboreType typeWellbore;
     @XmlSchemaType(name = "string")
     protected WellboreShape shape;
-    @XmlSchemaType(name = "dateTime")
-    protected XMLGregorianCalendar dTimKickoff;
-    protected Boolean achievedTD;
     protected MeasuredDepthCoord md;
     protected WellVerticalDepthCoord tvd;
     protected MeasuredDepthCoord mdBit;
@@ -106,58 +92,6 @@ public class ObjWellbore {
     protected TimeMeasure dayTarget;
     protected CsCommonData commonData;
     protected CsCustomData customData;
-    @XmlAttribute(name = "uidWell")
-    protected String uidWell;
-    @XmlAttribute(name = "uid")
-    protected String uid;
-
-    /**
-     * Gets the value of the nameWell property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getNameWell() {
-        return nameWell;
-    }
-
-    /**
-     * Sets the value of the nameWell property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setNameWell(String value) {
-        this.nameWell = value;
-    }
-
-    /**
-     * Gets the value of the name property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getName() {
-        return name;
-    }
-
-    /**
-     * Sets the value of the name property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setName(String value) {
-        this.name = value;
-    }
 
     /**
      * Gets the value of the parentWellbore property.
@@ -181,78 +115,6 @@ public class ObjWellbore {
      */
     public void setParentWellbore(RefNameString value) {
         this.parentWellbore = value;
-    }
-
-    /**
-     * Gets the value of the number property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getNumber() {
-        return number;
-    }
-
-    /**
-     * Sets the value of the number property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setNumber(String value) {
-        this.number = value;
-    }
-
-    /**
-     * Gets the value of the suffixAPI property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getSuffixAPI() {
-        return suffixAPI;
-    }
-
-    /**
-     * Sets the value of the suffixAPI property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setSuffixAPI(String value) {
-        this.suffixAPI = value;
-    }
-
-    /**
-     * Gets the value of the numGovt property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getNumGovt() {
-        return numGovt;
-    }
-
-    /**
-     * Sets the value of the numGovt property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setNumGovt(String value) {
-        this.numGovt = value;
     }
 
     /**
@@ -373,54 +235,6 @@ public class ObjWellbore {
      */
     public void setShape(WellboreShape value) {
         this.shape = value;
-    }
-
-    /**
-     * Gets the value of the dTimKickoff property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link XMLGregorianCalendar }
-     *     
-     */
-    public XMLGregorianCalendar getDTimKickoff() {
-        return dTimKickoff;
-    }
-
-    /**
-     * Sets the value of the dTimKickoff property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link XMLGregorianCalendar }
-     *     
-     */
-    public void setDTimKickoff(XMLGregorianCalendar value) {
-        this.dTimKickoff = value;
-    }
-
-    /**
-     * Gets the value of the achievedTD property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link Boolean }
-     *     
-     */
-    public Boolean isAchievedTD() {
-        return achievedTD;
-    }
-
-    /**
-     * Sets the value of the achievedTD property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link Boolean }
-     *     
-     */
-    public void setAchievedTD(Boolean value) {
-        this.achievedTD = value;
     }
 
     /**
@@ -734,53 +548,4 @@ public class ObjWellbore {
     public void setCustomData(CsCustomData value) {
         this.customData = value;
     }
-
-    /**
-     * Gets the value of the uidWell property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getUidWell() {
-        return uidWell;
-    }
-
-    /**
-     * Sets the value of the uidWell property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setUidWell(String value) {
-        this.uidWell = value;
-    }
-
-    /**
-     * Gets the value of the uid property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getUid() {
-        return uid;
-    }
-
-    /**
-     * Sets the value of the uid property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setUid(String value) {
-        this.uid = value;
-    }
-
 }
